Fix integer overflow in disk space calculation

+++ otsdaq/CoreSupervisors/CorePropertySupervisorBase.cc
@@ -186,7 +186,7 @@ void CorePropertySupervisorBase::getAvailableDiskSpace()
 		           << __E__;
 		__SUP_SS_THROW__;
 	}
-	uint64_t availableLogSpaceKB = (logStat.f_bavail * logStat.f_frsize) / 1024;  // in KB
+	uint64_t availableLogSpaceKB = (uint64_t)logStat.f_bavail * logStat.f_frsize / 1024;  // in KB
 	__SUP_COUTTV__(availableLogSpaceKB);
 
 	struct statvfs dataStat;
@@ -197,7 +197,7 @@ void CorePropertySupervisorBase::getAvailableDiskSpace()
 		__SUP_SS_THROW__;
 	}
 	uint64_t availableDataSpaceKB =
-	    (dataStat.f_bavail * dataStat.f_frsize) / 1024;  // in KB
+	    (uint64_t)dataStat.f_bavail * dataStat.f_frsize / 1024;  // in KB
 	__SUP_COUTTV__(availableDataSpaceKB);
 
 	availableLogSpaceKB_  = availableLogSpaceKB;
